examples/integrations/gcp/custom_example/custom_etl/requirements.txt
google-cloud-storage
google-cloud-secret-manager
labelbox
